Add Annotations, Add 4G compatability
The new settings view will show a slider if the
iphone as 4G capabilities. The maximum speed is
ajdusted depending on the model of iphone the app
is run on. The switch is required incase the person
doen't have a 4G contract.

An annoatation appear to tell the user the resulting
speed when an overlay is tapped.